requirements.txt
yt-dlp~=2025.2.19
python-multipart>=0.0.7
yt-dlp
scikit-learn~=1.6.1
pymongo==3.11
boto3~=1.37.4
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi~=0.111.0
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
idna==3.4
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.5
uvicorn>=0.12.0
pydantic>=2.10.0
annotated-types>=0.6.0
typing_extensions>=4.12.2
google~=3.0.0
botocore~=1.37.4
tqdm~=4.67.1
ytnoti
python-dotenv~=1.0.1
protobuf~=5.29.3
faster-whisper~=1.1.1
google.generativeai
browser-cookie3
selectorlib
bs4
git+https://github.com/coqui-ai/TTS.git